requirements.txt
torch==2.0.1
torchvision==0.15.2
pillow
numpy
google-generativeai>=0.3.0
requests
aiohttp
python-dotenv 